Adding cross domain frame busting and reverse-framebusting (to ensure bill text is /always/ displayed in a frame in a browser) in an SEO friendly way.
This is so that if pages are indexed by search engine site crawlers that readers end up in the right place.

This is a consequences of having to rely on the parsing of sometimes broken HTML that cannot be fixed automatically by HTML tidy libraries, which is why bill text is - regrettably - displayed in a frame in the first place.

It's be possible to strip bills back to plain text and re-display them is preferable, but such a parser would need to be able to cope with preserving tables while correctly identifying and stripping out junk and accounting for broken HTML.