New way of presenting bill text. New API URLs. New URLs for HTML and plain text versions of bill text. FAQ updated. Sitemap improved. Bugfixes.
Plain text and tidied HTML are now stored in the DB when bill text is scraped.

The API works by appending .json (or .html or .text where supported) to the end of a URL to get a version of a page in JSON or just the main content as HTML or Plain Text.

e.g.

For: http://public-scrutiny-office.org/bills
You can use: http://public-scrutiny-office.org/bills.json

For: http://public-scrutiny-office.org/bills/2013-2014/alan-turing-statutory-pardon/

You can use:
http://public-scrutiny-office.org/bills/2013-2014/alan-turing-statutory-pardon.json
http://public-scrutiny-office.org/bills/2013-2014/alan-turing-statutory-pardon.text
http://public-scrutiny-office.org/bills/2013-2014/alan-turing-statutory-pardon.html

This is similar to how Government Digital Service URLs work. I don't expect I'll be changing these URLs again any time soon so they are safe to use.

The content of of the JSON and text versions may vary slightly from the page as normally viewed in a browser, depending on context.