Updates for MongoDB being crummy.
Upgraded mongo on live to match the (newer) version on dev but then it broke for queries in a new and exciting way.

I guess the mongo folks couldn't decided how to handle sort order bool, and decided not to support a bool after all and go with a signed int. *slow clap*

Despite being the same version on the same data, it was differently for the same queries. Possibly an OS specific bug.

Anyway, it now works on both live and dev.