Now uses Parliamentary member ID for member ID objects to prevent duplicates.
Not sure how to handle elections/by elections yet.

May just write a script which runs monthly to check the status of all members and update a status field to be "inactive" if they are no longer serving and have the getMembers() and getMembersByName() methods only look for records that are still active.