Performance improvements and better mobile compatibility.
CSS improvements for readability but also performance, e.g. wrapping dynamic elements in containers of a fixed minimum height to reduce page render time.

Reduced the number of calls made on the homepage by NOT querying for how many comments there are for each post.

Button styles for the Approve/Disapprove buttons are now more similar to the default button style.

Added a link to jump to comments (to avoid having to scroll down through long bills).