Major maintenance - Bug fixes and CSS changes
This is a lot of fixes that paper over some cracks in the codebase as hotfixes to get it working again by removing borken features and cleaning up the UI a little to make it a bit more presentable.

There is a lot of terrible cruft here, and it doesn't use any of the new librarires than have been spun off from the code (which are much cleaner, have CI tests, etc), but this work is still an improvement.

There will be a rebuild soon which will focus on creating a new site with an editorial focus replace most of the code and update or replace all the dependancies.